Preserve log file time during log file rotation operation
When a file is compressed, its time was updated since a new compressed
file was created. Preserve the original time of the file so that
directory listings will show the time of the backup.

Added instructions in `README.md` on how to schedule `duplicacy-util`
jobs to run automatically on Linux, Mac OS/X, and Windows.