Added infrastructure for checkpointing duplicacy operations
Added infrastructure to allow for checkpointing of duplicacy
operations (continuing where last operation left off). This
is useful in case of a duplicacy failure, or network failure,
or a machine crash during a backup operation.

Note that checkpointing isn't integrated with the remainder
of the code yet. This commit includes utility functions with
unit tests.