Clean up error handling when a backup is already running
When a backup is running, a somewhat cryptic error was displayed,
and a mail message was sent as a failure. However, this can be a
normal situation depending on scheduling of backups.

The backup log now has: "Backup already running and will be skipped",
and the email subject ends with "(skipped)" rather than "(FAILURE)",
making it easier to filter on the messages if desired.

This fixes issue #17.