Modified handling of optional parameters in backup configuration
Previously, optional parmeters from backup configuraitons were
pre-parsed and always specified during duplicacy invocation
(even if not specified in backup configuration file). This became
problematic in duplicacy v2.1.1, where the prune command accepted
a (new) threads parameter.

Now handle all optional parameters optionally (specified on
command line only if speified in backup configuration file).
This allows the format of the backup configuration file to
control the version of duplicacy that is used, and not be tied
to the version of duplicacy-utils.